[v7] Remove `BTPayPalVaultBaseRequest` (#1462)
* remove BTPayPalVaultBaseRequest
* cleanup BTPayPalVaultRequest to subclass BTPayPalRequest
